eap/requirements.txt
elasticsearch==6.4.0
elasticsearch-dsl==6.4.0
python-dateutil==2.8.1
six==1.14.0
urllib3==1.25.8
